Split endpoinds by file
